walrus

Module ir

Intermediate representation for instructions.

The goal is to match wasm instructions as closely as possible, but translate the stack machine into an instruction tree. Additionally all control frames are representd as Blocks.

Structs

AnyConvertExtern

any.convert_extern - convert externref to anyref

ArrayCopy

array.copy - copy elements between arrays

ArrayFill

array.fill - fill a range of array elements with a value

ArrayGet

array.get - read an element from an array

ArrayGetS

array.get_s - read a packed element with sign extension

ArrayGetU

array.get_u - read a packed element with zero extension

ArrayInitData

array.init_data - initialize array elements from a data segment

ArrayInitElem

array.init_elem - initialize array elements from an element segment

ArrayLen

array.len - get the length of an array

ArrayNew

array.new - create a new array filled with a value

ArrayNewData

array.new_data - create a new array from a data segment

ArrayNewDefault

array.new_default - create a new array with default values

ArrayNewElem

array.new_elem - create a new array from an element segment

ArrayNewFixed

array.new_fixed - create a new array from stack values

ArraySet

array.set - write an element in an array

AtomicFence

The atomic.fence instruction

AtomicNotify

The atomic.notify instruction to wake up threads.

AtomicRmw

An atomic read/modify/write operation.

AtomicWait

The *.atomic.wait instruction to block threads.

Binop

Binary operations, those requiring two operands

Block

block ... end

Br

br

BrIf

br_if

BrOnCast

br_on_cast - branch if cast succeeds

BrOnCastFail

br_on_cast_fail - branch if cast fails

BrOnNonNull

br_on_non_null - branch if reference is non-null

BrOnNull

br_on_null - branch if reference is null

BrTable

br_table

Call

call

CallIndirect

call_indirect

CallRef

call_ref - call through a typed function reference

Cmpxchg

An atomic compare-and-exchange operation.

Const

*.const

DataDrop

data.drop

Drop

drop

ElemDrop

elem.drop

ExternConvertAny

extern.convert_any - convert anyref to externref

GlobalGet

global.get n

GlobalSet

global.set n

I8x16Shuffle

i8x16.shuffle

I8x16Swizzle

i8x16.swizzle

I31GetS

i31.get_s - extract signed i32 from i31ref

I31GetU

i31.get_u - extract unsigned i32 from i31ref

I64Add128

i64.add128 - 128-bit addition

I64MulWideS

i64.mul_wide_s - signed widening multiplication

I64MulWideU

i64.mul_wide_u - unsigned widening multiplication

I64Sub128

i64.sub128 - 128-bit subtraction

IfElse

if <consequent> else <alternative> end

InstrLocId

A symbolic original wasm operator source location.

InstrSeq

A sequence of instructions.

Load

*.load

LoadSimd

Various instructions to load a simd vector from memory

Local

A local variable or parameter.

LocalGet

local.get n

LocalSet

local.set n

LocalTee

local.tee n

Loop

loop ... end

MemArg

Arguments to memory operations, containing a constant offset from a dynamic address as well as a predicted alignment.

MemoryCopy

memory.copy

MemoryFill

memory.fill

MemoryGrow

memory.grow

MemoryInit

memory.init

MemorySize

memory.size

RefAsNonNull

ref.as_non_null - assert reference is non-null or trap

RefCast

ref.cast - cast reference to heap type, trap on failure

RefEq

ref.eq - compare two references for equality

RefFunc

ref.func

RefI31

ref.i31 - wrap i32 as i31ref

RefIsNull

ref.is_null

RefNull

ref.null $ty

RefTest

ref.test - test if reference matches heap type

Rethrow

rethrow relative_depth - rethrow a caught exception (legacy)

Return

return

ReturnCall

return_call

ReturnCallIndirect

return_call_indirect

ReturnCallRef

return_call_ref - tail call through a typed function reference

Select

select

Store

*.store

StructGet

struct.get - read a field from a struct

StructGetS

struct.get_s - read a packed field with sign extension

StructGetU

struct.get_u - read a packed field with zero extension

StructNew

struct.new - create a new struct instance

StructNewDefault

struct.new_default - create a new struct with default field values

StructSet

struct.set - write a field in a struct

TableCopy

table.copy

TableFill

table.fill

TableGet

table.get

TableGrow

table.grow

TableInit

table.init

TableSet

table.set

TableSize

table.size

TernOp

Ternary operations, those requiring three operands

Throw

throw - throw an exception

ThrowRef

throw_ref - rethrow a caught exception reference

Try

try blocktype ... end - try block with catch handlers (legacy)

TryTable

try_table ... end - exception handling with catch table (modern proposal)

Unop

Unary operations, those requiring one operand

Unreachable

unreachable

V128Bitselect

v128.bitselect

Enums

AtomicOp

The different kinds of atomic rmw operations

AtomicWidth

The different kinds of atomic rmw operations

BinaryOp

Possible binary operations in wasm

ExtendedLoad

The kinds of extended loads which can happen

Instr

An enum of all the different kinds of wasm instructions.

InstrSeqType

The type of an instruction sequence.

LegacyCatch

A catch clause in a legacy Try instruction

LoadKind

The different kinds of load instructions that are part of a Load IR node

LoadSimdKind

The different kinds of load instructions that are part of a LoadSimd IR node

StoreKind

The different kinds of store instructions that are part of a Store IR node

TernaryOp

Possible ternary operations in wasm

TryTableCatch

A catch clause in a TryTable instruction

UnaryOp

Possible unary operations in wasm

Value

Constant values that can show up in WebAssembly

Traits

Visitor

A visitor is a set of callbacks that are called when a traversal (such as dfs_in_order) is walking an instruction tree.

VisitorMut

A mutable version of Visitor.

Functions

dfs_in_order

Perform an intra-procedural, depth-first, in-order traversal of the IR.

dfs_pre_order_mut

Perform an intra-procedural, depth-first, pre-order, mutable traversal of the IR.

Type Aliases

InstrSeqId

The identifier for a InstrSeq within some LocalFunction.

LocalId

The id of a local.

ShuffleIndices

Argument in V128Shuffle of lane indices to select