walrus::ir

Trait Visitor

pub trait Visitor<'instr>: Sized {

    // Provided methods
    fn start_instr_seq(&mut self, instr_seq: &'instr InstrSeq) { ... }
    fn end_instr_seq(&mut self, instr_seq: &'instr InstrSeq) { ... }
    fn visit_instr(
        &mut self,
        instr: &'instr Instr,
        instr_loc: &'instr InstrLocId,
    ) { ... }
    fn visit_instr_seq_id(&mut self, instr_seq_id: &InstrSeqId) { ... }
    fn visit_local_id(&mut self, local: &LocalId) { ... }
    fn visit_memory_id(&mut self, memory: &MemoryId) { ... }
    fn visit_table_id(&mut self, table: &TableId) { ... }
    fn visit_global_id(&mut self, global: &GlobalId) { ... }
    fn visit_function_id(&mut self, function: &FunctionId) { ... }
    fn visit_data_id(&mut self, function: &DataId) { ... }
    fn visit_type_id(&mut self, ty: &TypeId) { ... }
    fn visit_element_id(&mut self, elem: &ElementId) { ... }
    fn visit_tag_id(&mut self, tag: &TagId) { ... }
    fn visit_value(&mut self, value: &Value) { ... }
    fn visit_heap_type(&mut self, heap_type: &HeapType) { ... }
    fn visit_block(&mut self, instr: &Block) { ... }
    fn visit_loop(&mut self, instr: &Loop) { ... }
    fn visit_call(&mut self, instr: &Call) { ... }
    fn visit_call_indirect(&mut self, instr: &CallIndirect) { ... }
    fn visit_local_get(&mut self, instr: &LocalGet) { ... }
    fn visit_local_set(&mut self, instr: &LocalSet) { ... }
    fn visit_local_tee(&mut self, instr: &LocalTee) { ... }
    fn visit_global_get(&mut self, instr: &GlobalGet) { ... }
    fn visit_global_set(&mut self, instr: &GlobalSet) { ... }
    fn visit_const(&mut self, instr: &Const) { ... }
    fn visit_tern_op(&mut self, instr: &TernOp) { ... }
    fn visit_binop(&mut self, instr: &Binop) { ... }
    fn visit_unop(&mut self, instr: &Unop) { ... }
    fn visit_select(&mut self, instr: &Select) { ... }
    fn visit_unreachable(&mut self, instr: &Unreachable) { ... }
    fn visit_br(&mut self, instr: &Br) { ... }
    fn visit_br_if(&mut self, instr: &BrIf) { ... }
    fn visit_if_else(&mut self, instr: &IfElse) { ... }
    fn visit_br_table(&mut self, instr: &BrTable) { ... }
    fn visit_drop(&mut self, instr: &Drop) { ... }
    fn visit_return(&mut self, instr: &Return) { ... }
    fn visit_memory_size(&mut self, instr: &MemorySize) { ... }
    fn visit_memory_grow(&mut self, instr: &MemoryGrow) { ... }
    fn visit_memory_init(&mut self, instr: &MemoryInit) { ... }
    fn visit_data_drop(&mut self, instr: &DataDrop) { ... }
    fn visit_memory_copy(&mut self, instr: &MemoryCopy) { ... }
    fn visit_memory_fill(&mut self, instr: &MemoryFill) { ... }
    fn visit_load(&mut self, instr: &Load) { ... }
    fn visit_store(&mut self, instr: &Store) { ... }
    fn visit_atomic_rmw(&mut self, instr: &AtomicRmw) { ... }
    fn visit_cmpxchg(&mut self, instr: &Cmpxchg) { ... }
    fn visit_atomic_notify(&mut self, instr: &AtomicNotify) { ... }
    fn visit_atomic_wait(&mut self, instr: &AtomicWait) { ... }
    fn visit_atomic_fence(&mut self, instr: &AtomicFence) { ... }
    fn visit_table_get(&mut self, instr: &TableGet) { ... }
    fn visit_table_set(&mut self, instr: &TableSet) { ... }
    fn visit_table_grow(&mut self, instr: &TableGrow) { ... }
    fn visit_table_size(&mut self, instr: &TableSize) { ... }
    fn visit_table_fill(&mut self, instr: &TableFill) { ... }
    fn visit_ref_null(&mut self, instr: &RefNull) { ... }
    fn visit_ref_is_null(&mut self, instr: &RefIsNull) { ... }
    fn visit_ref_func(&mut self, instr: &RefFunc) { ... }
    fn visit_ref_as_non_null(&mut self, instr: &RefAsNonNull) { ... }
    fn visit_br_on_null(&mut self, instr: &BrOnNull) { ... }
    fn visit_br_on_non_null(&mut self, instr: &BrOnNonNull) { ... }
    fn visit_call_ref(&mut self, instr: &CallRef) { ... }
    fn visit_return_call_ref(&mut self, instr: &ReturnCallRef) { ... }
    fn visit_ref_i31(&mut self, instr: &RefI31) { ... }
    fn visit_i31_get_s(&mut self, instr: &I31GetS) { ... }
    fn visit_i31_get_u(&mut self, instr: &I31GetU) { ... }
    fn visit_ref_test(&mut self, instr: &RefTest) { ... }
    fn visit_ref_cast(&mut self, instr: &RefCast) { ... }
    fn visit_br_on_cast(&mut self, instr: &BrOnCast) { ... }
    fn visit_br_on_cast_fail(&mut self, instr: &BrOnCastFail) { ... }
    fn visit_any_convert_extern(&mut self, instr: &AnyConvertExtern) { ... }
    fn visit_extern_convert_any(&mut self, instr: &ExternConvertAny) { ... }
    fn visit_ref_eq(&mut self, instr: &RefEq) { ... }
    fn visit_struct_new(&mut self, instr: &StructNew) { ... }
    fn visit_struct_new_default(&mut self, instr: &StructNewDefault) { ... }
    fn visit_struct_get(&mut self, instr: &StructGet) { ... }
    fn visit_struct_get_s(&mut self, instr: &StructGetS) { ... }
    fn visit_struct_get_u(&mut self, instr: &StructGetU) { ... }
    fn visit_struct_set(&mut self, instr: &StructSet) { ... }
    fn visit_array_new(&mut self, instr: &ArrayNew) { ... }
    fn visit_array_new_default(&mut self, instr: &ArrayNewDefault) { ... }
    fn visit_array_new_fixed(&mut self, instr: &ArrayNewFixed) { ... }
    fn visit_array_new_data(&mut self, instr: &ArrayNewData) { ... }
    fn visit_array_new_elem(&mut self, instr: &ArrayNewElem) { ... }
    fn visit_array_get(&mut self, instr: &ArrayGet) { ... }
    fn visit_array_get_s(&mut self, instr: &ArrayGetS) { ... }
    fn visit_array_get_u(&mut self, instr: &ArrayGetU) { ... }
    fn visit_array_set(&mut self, instr: &ArraySet) { ... }
    fn visit_array_len(&mut self, instr: &ArrayLen) { ... }
    fn visit_array_fill(&mut self, instr: &ArrayFill) { ... }
    fn visit_array_copy(&mut self, instr: &ArrayCopy) { ... }
    fn visit_array_init_data(&mut self, instr: &ArrayInitData) { ... }
    fn visit_array_init_elem(&mut self, instr: &ArrayInitElem) { ... }
    fn visit_i64_add128(&mut self, instr: &I64Add128) { ... }
    fn visit_i64_sub128(&mut self, instr: &I64Sub128) { ... }
    fn visit_i64_mul_wide_s(&mut self, instr: &I64MulWideS) { ... }
    fn visit_i64_mul_wide_u(&mut self, instr: &I64MulWideU) { ... }
    fn visit_v128_bitselect(&mut self, instr: &V128Bitselect) { ... }
    fn visit_i8x16_swizzle(&mut self, instr: &I8x16Swizzle) { ... }
    fn visit_i8x16_shuffle(&mut self, instr: &I8x16Shuffle) { ... }
    fn visit_load_simd(&mut self, instr: &LoadSimd) { ... }
    fn visit_table_init(&mut self, instr: &TableInit) { ... }
    fn visit_elem_drop(&mut self, instr: &ElemDrop) { ... }
    fn visit_table_copy(&mut self, instr: &TableCopy) { ... }
    fn visit_return_call(&mut self, instr: &ReturnCall) { ... }
    fn visit_return_call_indirect(&mut self, instr: &ReturnCallIndirect) { ... }
    fn visit_try_table(&mut self, instr: &TryTable) { ... }
    fn visit_throw(&mut self, instr: &Throw) { ... }
    fn visit_throw_ref(&mut self, instr: &ThrowRef) { ... }
    fn visit_try(&mut self, instr: &Try) { ... }
    fn visit_rethrow(&mut self, instr: &Rethrow) { ... }

}

A visitor is a set of callbacks that are called when a traversal (such as dfs_in_order) is walking an instruction tree.

Recursion

Do not recursively get nested InstrSeqs for any InstrSeqId you visit! You will blow the stack when processing large Wasm files. Visitors are just heterogenously-typed callbacks, not traversals themselves!

Instead, use walrus::ir::dfs_in_order and other traversal drivers that will walk the tree in a non-recursive, iterative fashion

Provided Methods

Every Visitor trait method has a default, provided implementation that does nothing.

Provided Methods

fn start_instr_seq(&mut self, instr_seq: &'instr InstrSeq)

Called before the traversal will start visiting each of the instructions an instruction sequence.

The order in which instruction sequences are visited is defined by the traversal function, e.g. walrus::ir::dfs_in_order.

fn end_instr_seq(&mut self, instr_seq: &'instr InstrSeq)

Called after the traversal finishes visiting each of the instructions in an instruction sequence.

fn visit_instr(&mut self, instr: &'instr Instr, instr_loc: &'instr InstrLocId)

Visit Instr.

fn visit_instr_seq_id(&mut self, instr_seq_id: &InstrSeqId)

Visit InstrSeqId.

fn visit_local_id(&mut self, local: &LocalId)

Visit LocalId.

fn visit_memory_id(&mut self, memory: &MemoryId)

Visit MemoryId.

fn visit_table_id(&mut self, table: &TableId)

Visit TableId.

fn visit_global_id(&mut self, global: &GlobalId)

Visit GlobalId.

fn visit_function_id(&mut self, function: &FunctionId)

Visit FunctionId.

fn visit_data_id(&mut self, function: &DataId)

Visit DataId.

fn visit_type_id(&mut self, ty: &TypeId)

Visit TypeId

fn visit_element_id(&mut self, elem: &ElementId)

Visit ElementId

fn visit_tag_id(&mut self, tag: &TagId)

Visit TagId

fn visit_value(&mut self, value: &Value)

Visit Value.

fn visit_heap_type(&mut self, heap_type: &HeapType)

Visit HeapType.

The default implementation extracts the TypeId from HeapType::Concrete or HeapType::Exact and forwards it to visit_type_id. Override this to handle heap types directly (e.g., to distinguish concrete from exact).

fn visit_block(&mut self, instr: &Block)

Visit Block.

fn visit_loop(&mut self, instr: &Loop)

Visit Loop.

fn visit_call(&mut self, instr: &Call)

Visit Call.

fn visit_call_indirect(&mut self, instr: &CallIndirect)

Visit CallIndirect.

fn visit_local_get(&mut self, instr: &LocalGet)

Visit LocalGet.

fn visit_local_set(&mut self, instr: &LocalSet)

Visit LocalSet.

fn visit_local_tee(&mut self, instr: &LocalTee)

Visit LocalTee.

fn visit_global_get(&mut self, instr: &GlobalGet)

Visit GlobalGet.

fn visit_global_set(&mut self, instr: &GlobalSet)

Visit GlobalSet.

fn visit_const(&mut self, instr: &Const)

Visit Const.

fn visit_tern_op(&mut self, instr: &TernOp)

Visit TernOp.

fn visit_binop(&mut self, instr: &Binop)

Visit Binop.

fn visit_unop(&mut self, instr: &Unop)

Visit Unop.

fn visit_select(&mut self, instr: &Select)

Visit Select.

fn visit_unreachable(&mut self, instr: &Unreachable)

Visit Unreachable.

fn visit_br(&mut self, instr: &Br)

Visit Br.

fn visit_br_if(&mut self, instr: &BrIf)

Visit BrIf.

fn visit_if_else(&mut self, instr: &IfElse)

Visit IfElse.

fn visit_br_table(&mut self, instr: &BrTable)

Visit BrTable.

fn visit_drop(&mut self, instr: &Drop)

Visit Drop.

fn visit_return(&mut self, instr: &Return)

Visit Return.

fn visit_memory_size(&mut self, instr: &MemorySize)

Visit MemorySize.

fn visit_memory_grow(&mut self, instr: &MemoryGrow)

Visit MemoryGrow.

fn visit_memory_init(&mut self, instr: &MemoryInit)

Visit MemoryInit.

fn visit_data_drop(&mut self, instr: &DataDrop)

Visit DataDrop.

fn visit_memory_copy(&mut self, instr: &MemoryCopy)

Visit MemoryCopy.

fn visit_memory_fill(&mut self, instr: &MemoryFill)

Visit MemoryFill.

fn visit_load(&mut self, instr: &Load)

Visit Load.

fn visit_store(&mut self, instr: &Store)

Visit Store.

fn visit_atomic_rmw(&mut self, instr: &AtomicRmw)

Visit AtomicRmw.

fn visit_cmpxchg(&mut self, instr: &Cmpxchg)

Visit Cmpxchg.

fn visit_atomic_notify(&mut self, instr: &AtomicNotify)

Visit AtomicNotify.

fn visit_atomic_wait(&mut self, instr: &AtomicWait)

Visit AtomicWait.

fn visit_atomic_fence(&mut self, instr: &AtomicFence)

Visit AtomicFence.

fn visit_table_get(&mut self, instr: &TableGet)

Visit TableGet.

fn visit_table_set(&mut self, instr: &TableSet)

Visit TableSet.

fn visit_table_grow(&mut self, instr: &TableGrow)

Visit TableGrow.

fn visit_table_size(&mut self, instr: &TableSize)

Visit TableSize.

fn visit_table_fill(&mut self, instr: &TableFill)

Visit TableFill.

fn visit_ref_null(&mut self, instr: &RefNull)

Visit RefNull.

fn visit_ref_is_null(&mut self, instr: &RefIsNull)

Visit RefIsNull.

fn visit_ref_func(&mut self, instr: &RefFunc)

Visit RefFunc.

fn visit_ref_as_non_null(&mut self, instr: &RefAsNonNull)

Visit RefAsNonNull.

fn visit_br_on_null(&mut self, instr: &BrOnNull)

Visit BrOnNull.

fn visit_br_on_non_null(&mut self, instr: &BrOnNonNull)

Visit BrOnNonNull.

fn visit_call_ref(&mut self, instr: &CallRef)

Visit CallRef.

fn visit_return_call_ref(&mut self, instr: &ReturnCallRef)

Visit ReturnCallRef.

fn visit_ref_i31(&mut self, instr: &RefI31)

Visit RefI31.

fn visit_i31_get_s(&mut self, instr: &I31GetS)

Visit I31GetS.

fn visit_i31_get_u(&mut self, instr: &I31GetU)

Visit I31GetU.

fn visit_ref_test(&mut self, instr: &RefTest)

Visit RefTest.

fn visit_ref_cast(&mut self, instr: &RefCast)

Visit RefCast.

fn visit_br_on_cast(&mut self, instr: &BrOnCast)

Visit BrOnCast.

fn visit_br_on_cast_fail(&mut self, instr: &BrOnCastFail)

Visit BrOnCastFail.

fn visit_any_convert_extern(&mut self, instr: &AnyConvertExtern)

Visit AnyConvertExtern.

fn visit_extern_convert_any(&mut self, instr: &ExternConvertAny)

Visit ExternConvertAny.

fn visit_ref_eq(&mut self, instr: &RefEq)

Visit RefEq.

fn visit_struct_new(&mut self, instr: &StructNew)

Visit StructNew.

fn visit_struct_new_default(&mut self, instr: &StructNewDefault)

Visit StructNewDefault.

fn visit_struct_get(&mut self, instr: &StructGet)

Visit StructGet.

fn visit_struct_get_s(&mut self, instr: &StructGetS)

Visit StructGetS.

fn visit_struct_get_u(&mut self, instr: &StructGetU)

Visit StructGetU.

fn visit_struct_set(&mut self, instr: &StructSet)

Visit StructSet.

fn visit_array_new(&mut self, instr: &ArrayNew)

Visit ArrayNew.

fn visit_array_new_default(&mut self, instr: &ArrayNewDefault)

Visit ArrayNewDefault.

fn visit_array_new_fixed(&mut self, instr: &ArrayNewFixed)

Visit ArrayNewFixed.

fn visit_array_new_data(&mut self, instr: &ArrayNewData)

Visit ArrayNewData.

fn visit_array_new_elem(&mut self, instr: &ArrayNewElem)

Visit ArrayNewElem.

fn visit_array_get(&mut self, instr: &ArrayGet)

Visit ArrayGet.

fn visit_array_get_s(&mut self, instr: &ArrayGetS)

Visit ArrayGetS.

fn visit_array_get_u(&mut self, instr: &ArrayGetU)

Visit ArrayGetU.

fn visit_array_set(&mut self, instr: &ArraySet)

Visit ArraySet.

fn visit_array_len(&mut self, instr: &ArrayLen)

Visit ArrayLen.

fn visit_array_fill(&mut self, instr: &ArrayFill)

Visit ArrayFill.

fn visit_array_copy(&mut self, instr: &ArrayCopy)

Visit ArrayCopy.

fn visit_array_init_data(&mut self, instr: &ArrayInitData)

Visit ArrayInitData.

fn visit_array_init_elem(&mut self, instr: &ArrayInitElem)

Visit ArrayInitElem.

fn visit_i64_add128(&mut self, instr: &I64Add128)

Visit I64Add128.

fn visit_i64_sub128(&mut self, instr: &I64Sub128)

Visit I64Sub128.

fn visit_i64_mul_wide_s(&mut self, instr: &I64MulWideS)

Visit I64MulWideS.

fn visit_i64_mul_wide_u(&mut self, instr: &I64MulWideU)

Visit I64MulWideU.

fn visit_v128_bitselect(&mut self, instr: &V128Bitselect)

Visit V128Bitselect.

fn visit_i8x16_swizzle(&mut self, instr: &I8x16Swizzle)

Visit I8x16Swizzle.

fn visit_i8x16_shuffle(&mut self, instr: &I8x16Shuffle)

Visit I8x16Shuffle.

fn visit_load_simd(&mut self, instr: &LoadSimd)

Visit LoadSimd.

fn visit_table_init(&mut self, instr: &TableInit)

Visit TableInit.

fn visit_elem_drop(&mut self, instr: &ElemDrop)

Visit ElemDrop.

fn visit_table_copy(&mut self, instr: &TableCopy)

Visit TableCopy.

fn visit_return_call(&mut self, instr: &ReturnCall)

Visit ReturnCall.

fn visit_return_call_indirect(&mut self, instr: &ReturnCallIndirect)

Visit ReturnCallIndirect.

fn visit_try_table(&mut self, instr: &TryTable)

Visit TryTable.

fn visit_throw(&mut self, instr: &Throw)

Visit Throw.

fn visit_throw_ref(&mut self, instr: &ThrowRef)

Visit ThrowRef.

fn visit_try(&mut self, instr: &Try)

Visit Try.

fn visit_rethrow(&mut self, instr: &Rethrow)

Visit Rethrow.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.