walrus::ir

Trait VisitorMut

pub trait VisitorMut: Sized {

    // Provided methods
    fn start_instr_seq_mut(&mut self, instr_seq: &mut InstrSeq) { ... }
    fn end_instr_seq_mut(&mut self, instr_seq: &mut InstrSeq) { ... }
    fn visit_instr_mut(&mut self, instr: &mut Instr, instr_loc: &mut InstrLocId) { ... }
    fn visit_instr_seq_id_mut(&mut self, instr_seq_id: &mut InstrSeqId) { ... }
    fn visit_local_id_mut(&mut self, local: &mut LocalId) { ... }
    fn visit_memory_id_mut(&mut self, memory: &mut MemoryId) { ... }
    fn visit_table_id_mut(&mut self, table: &mut TableId) { ... }
    fn visit_global_id_mut(&mut self, global: &mut GlobalId) { ... }
    fn visit_function_id_mut(&mut self, function: &mut FunctionId) { ... }
    fn visit_data_id_mut(&mut self, function: &mut DataId) { ... }
    fn visit_type_id_mut(&mut self, ty: &mut TypeId) { ... }
    fn visit_element_id_mut(&mut self, elem: &mut ElementId) { ... }
    fn visit_tag_id_mut(&mut self, tag: &mut TagId) { ... }
    fn visit_value_mut(&mut self, value: &mut Value) { ... }
    fn visit_heap_type_mut(&mut self, heap_type: &mut HeapType) { ... }
    fn visit_block_mut(&mut self, instr: &mut Block) { ... }
    fn visit_loop_mut(&mut self, instr: &mut Loop) { ... }
    fn visit_call_mut(&mut self, instr: &mut Call) { ... }
    fn visit_call_indirect_mut(&mut self, instr: &mut CallIndirect) { ... }
    fn visit_local_get_mut(&mut self, instr: &mut LocalGet) { ... }
    fn visit_local_set_mut(&mut self, instr: &mut LocalSet) { ... }
    fn visit_local_tee_mut(&mut self, instr: &mut LocalTee) { ... }
    fn visit_global_get_mut(&mut self, instr: &mut GlobalGet) { ... }
    fn visit_global_set_mut(&mut self, instr: &mut GlobalSet) { ... }
    fn visit_const_mut(&mut self, instr: &mut Const) { ... }
    fn visit_tern_op_mut(&mut self, instr: &mut TernOp) { ... }
    fn visit_binop_mut(&mut self, instr: &mut Binop) { ... }
    fn visit_unop_mut(&mut self, instr: &mut Unop) { ... }
    fn visit_select_mut(&mut self, instr: &mut Select) { ... }
    fn visit_unreachable_mut(&mut self, instr: &mut Unreachable) { ... }
    fn visit_br_mut(&mut self, instr: &mut Br) { ... }
    fn visit_br_if_mut(&mut self, instr: &mut BrIf) { ... }
    fn visit_if_else_mut(&mut self, instr: &mut IfElse) { ... }
    fn visit_br_table_mut(&mut self, instr: &mut BrTable) { ... }
    fn visit_drop_mut(&mut self, instr: &mut Drop) { ... }
    fn visit_return_mut(&mut self, instr: &mut Return) { ... }
    fn visit_memory_size_mut(&mut self, instr: &mut MemorySize) { ... }
    fn visit_memory_grow_mut(&mut self, instr: &mut MemoryGrow) { ... }
    fn visit_memory_init_mut(&mut self, instr: &mut MemoryInit) { ... }
    fn visit_data_drop_mut(&mut self, instr: &mut DataDrop) { ... }
    fn visit_memory_copy_mut(&mut self, instr: &mut MemoryCopy) { ... }
    fn visit_memory_fill_mut(&mut self, instr: &mut MemoryFill) { ... }
    fn visit_load_mut(&mut self, instr: &mut Load) { ... }
    fn visit_store_mut(&mut self, instr: &mut Store) { ... }
    fn visit_atomic_rmw_mut(&mut self, instr: &mut AtomicRmw) { ... }
    fn visit_cmpxchg_mut(&mut self, instr: &mut Cmpxchg) { ... }
    fn visit_atomic_notify_mut(&mut self, instr: &mut AtomicNotify) { ... }
    fn visit_atomic_wait_mut(&mut self, instr: &mut AtomicWait) { ... }
    fn visit_atomic_fence_mut(&mut self, instr: &mut AtomicFence) { ... }
    fn visit_table_get_mut(&mut self, instr: &mut TableGet) { ... }
    fn visit_table_set_mut(&mut self, instr: &mut TableSet) { ... }
    fn visit_table_grow_mut(&mut self, instr: &mut TableGrow) { ... }
    fn visit_table_size_mut(&mut self, instr: &mut TableSize) { ... }
    fn visit_table_fill_mut(&mut self, instr: &mut TableFill) { ... }
    fn visit_ref_null_mut(&mut self, instr: &mut RefNull) { ... }
    fn visit_ref_is_null_mut(&mut self, instr: &mut RefIsNull) { ... }
    fn visit_ref_func_mut(&mut self, instr: &mut RefFunc) { ... }
    fn visit_ref_as_non_null_mut(&mut self, instr: &mut RefAsNonNull) { ... }
    fn visit_br_on_null_mut(&mut self, instr: &mut BrOnNull) { ... }
    fn visit_br_on_non_null_mut(&mut self, instr: &mut BrOnNonNull) { ... }
    fn visit_call_ref_mut(&mut self, instr: &mut CallRef) { ... }
    fn visit_return_call_ref_mut(&mut self, instr: &mut ReturnCallRef) { ... }
    fn visit_ref_i31_mut(&mut self, instr: &mut RefI31) { ... }
    fn visit_i31_get_s_mut(&mut self, instr: &mut I31GetS) { ... }
    fn visit_i31_get_u_mut(&mut self, instr: &mut I31GetU) { ... }
    fn visit_ref_test_mut(&mut self, instr: &mut RefTest) { ... }
    fn visit_ref_cast_mut(&mut self, instr: &mut RefCast) { ... }
    fn visit_br_on_cast_mut(&mut self, instr: &mut BrOnCast) { ... }
    fn visit_br_on_cast_fail_mut(&mut self, instr: &mut BrOnCastFail) { ... }
    fn visit_any_convert_extern_mut(&mut self, instr: &mut AnyConvertExtern) { ... }
    fn visit_extern_convert_any_mut(&mut self, instr: &mut ExternConvertAny) { ... }
    fn visit_ref_eq_mut(&mut self, instr: &mut RefEq) { ... }
    fn visit_struct_new_mut(&mut self, instr: &mut StructNew) { ... }
    fn visit_struct_new_default_mut(&mut self, instr: &mut StructNewDefault) { ... }
    fn visit_struct_get_mut(&mut self, instr: &mut StructGet) { ... }
    fn visit_struct_get_s_mut(&mut self, instr: &mut StructGetS) { ... }
    fn visit_struct_get_u_mut(&mut self, instr: &mut StructGetU) { ... }
    fn visit_struct_set_mut(&mut self, instr: &mut StructSet) { ... }
    fn visit_array_new_mut(&mut self, instr: &mut ArrayNew) { ... }
    fn visit_array_new_default_mut(&mut self, instr: &mut ArrayNewDefault) { ... }
    fn visit_array_new_fixed_mut(&mut self, instr: &mut ArrayNewFixed) { ... }
    fn visit_array_new_data_mut(&mut self, instr: &mut ArrayNewData) { ... }
    fn visit_array_new_elem_mut(&mut self, instr: &mut ArrayNewElem) { ... }
    fn visit_array_get_mut(&mut self, instr: &mut ArrayGet) { ... }
    fn visit_array_get_s_mut(&mut self, instr: &mut ArrayGetS) { ... }
    fn visit_array_get_u_mut(&mut self, instr: &mut ArrayGetU) { ... }
    fn visit_array_set_mut(&mut self, instr: &mut ArraySet) { ... }
    fn visit_array_len_mut(&mut self, instr: &mut ArrayLen) { ... }
    fn visit_array_fill_mut(&mut self, instr: &mut ArrayFill) { ... }
    fn visit_array_copy_mut(&mut self, instr: &mut ArrayCopy) { ... }
    fn visit_array_init_data_mut(&mut self, instr: &mut ArrayInitData) { ... }
    fn visit_array_init_elem_mut(&mut self, instr: &mut ArrayInitElem) { ... }
    fn visit_i64_add128_mut(&mut self, instr: &mut I64Add128) { ... }
    fn visit_i64_sub128_mut(&mut self, instr: &mut I64Sub128) { ... }
    fn visit_i64_mul_wide_s_mut(&mut self, instr: &mut I64MulWideS) { ... }
    fn visit_i64_mul_wide_u_mut(&mut self, instr: &mut I64MulWideU) { ... }
    fn visit_v128_bitselect_mut(&mut self, instr: &mut V128Bitselect) { ... }
    fn visit_i8x16_swizzle_mut(&mut self, instr: &mut I8x16Swizzle) { ... }
    fn visit_i8x16_shuffle_mut(&mut self, instr: &mut I8x16Shuffle) { ... }
    fn visit_load_simd_mut(&mut self, instr: &mut LoadSimd) { ... }
    fn visit_table_init_mut(&mut self, instr: &mut TableInit) { ... }
    fn visit_elem_drop_mut(&mut self, instr: &mut ElemDrop) { ... }
    fn visit_table_copy_mut(&mut self, instr: &mut TableCopy) { ... }
    fn visit_return_call_mut(&mut self, instr: &mut ReturnCall) { ... }
    fn visit_return_call_indirect_mut(&mut self, instr: &mut ReturnCallIndirect) { ... }
    fn visit_try_table_mut(&mut self, instr: &mut TryTable) { ... }
    fn visit_throw_mut(&mut self, instr: &mut Throw) { ... }
    fn visit_throw_ref_mut(&mut self, instr: &mut ThrowRef) { ... }
    fn visit_try_mut(&mut self, instr: &mut Try) { ... }
    fn visit_rethrow_mut(&mut self, instr: &mut Rethrow) { ... }

}

A mutable version of Visitor.

See Visitor’s documentation for details.

Provided Methods

fn start_instr_seq_mut(&mut self, instr_seq: &mut InstrSeq)

Called before the traversal will start visiting each of the instructions an instruction sequence.

The order in which instruction sequences are visited is defined by the traversal function, e.g. walrus::ir::dfs_pre_order_mut.

fn end_instr_seq_mut(&mut self, instr_seq: &mut InstrSeq)

Called after the traversal finishes visiting each of the instructions in an instruction sequence.

fn visit_instr_mut(&mut self, instr: &mut Instr, instr_loc: &mut InstrLocId)

Visit Instr.

fn visit_instr_seq_id_mut(&mut self, instr_seq_id: &mut InstrSeqId)

Visit InstrSeqId.

fn visit_local_id_mut(&mut self, local: &mut LocalId)

Visit Local.

fn visit_memory_id_mut(&mut self, memory: &mut MemoryId)

Visit Memory.

fn visit_table_id_mut(&mut self, table: &mut TableId)

Visit Table.

fn visit_global_id_mut(&mut self, global: &mut GlobalId)

Visit GlobalId.

fn visit_function_id_mut(&mut self, function: &mut FunctionId)

Visit FunctionId.

fn visit_data_id_mut(&mut self, function: &mut DataId)

Visit DataId.

fn visit_type_id_mut(&mut self, ty: &mut TypeId)

Visit TypeId

fn visit_element_id_mut(&mut self, elem: &mut ElementId)

Visit ElementId

fn visit_tag_id_mut(&mut self, tag: &mut TagId)

Visit TagId

fn visit_value_mut(&mut self, value: &mut Value)

Visit Value.

fn visit_heap_type_mut(&mut self, heap_type: &mut HeapType)

Visit HeapType mutably.

The default implementation extracts the TypeId from HeapType::Concrete or HeapType::Exact and forwards it to visit_type_id_mut. Override this to handle heap types directly.

fn visit_block_mut(&mut self, instr: &mut Block)

Visit Block.

fn visit_loop_mut(&mut self, instr: &mut Loop)

Visit Loop.

fn visit_call_mut(&mut self, instr: &mut Call)

Visit Call.

fn visit_call_indirect_mut(&mut self, instr: &mut CallIndirect)

Visit CallIndirect.

fn visit_local_get_mut(&mut self, instr: &mut LocalGet)

Visit LocalGet.

fn visit_local_set_mut(&mut self, instr: &mut LocalSet)

Visit LocalSet.

fn visit_local_tee_mut(&mut self, instr: &mut LocalTee)

Visit LocalTee.

fn visit_global_get_mut(&mut self, instr: &mut GlobalGet)

Visit GlobalGet.

fn visit_global_set_mut(&mut self, instr: &mut GlobalSet)

Visit GlobalSet.

fn visit_const_mut(&mut self, instr: &mut Const)

Visit Const.

fn visit_tern_op_mut(&mut self, instr: &mut TernOp)

Visit TernOp.

fn visit_binop_mut(&mut self, instr: &mut Binop)

Visit Binop.

fn visit_unop_mut(&mut self, instr: &mut Unop)

Visit Unop.

fn visit_select_mut(&mut self, instr: &mut Select)

Visit Select.

fn visit_unreachable_mut(&mut self, instr: &mut Unreachable)

Visit Unreachable.

fn visit_br_mut(&mut self, instr: &mut Br)

Visit Br.

fn visit_br_if_mut(&mut self, instr: &mut BrIf)

Visit BrIf.

fn visit_if_else_mut(&mut self, instr: &mut IfElse)

Visit IfElse.

fn visit_br_table_mut(&mut self, instr: &mut BrTable)

Visit BrTable.

fn visit_drop_mut(&mut self, instr: &mut Drop)

Visit Drop.

fn visit_return_mut(&mut self, instr: &mut Return)

Visit Return.

fn visit_memory_size_mut(&mut self, instr: &mut MemorySize)

Visit MemorySize.

fn visit_memory_grow_mut(&mut self, instr: &mut MemoryGrow)

Visit MemoryGrow.

fn visit_memory_init_mut(&mut self, instr: &mut MemoryInit)

Visit MemoryInit.

fn visit_data_drop_mut(&mut self, instr: &mut DataDrop)

Visit DataDrop.

fn visit_memory_copy_mut(&mut self, instr: &mut MemoryCopy)

Visit MemoryCopy.

fn visit_memory_fill_mut(&mut self, instr: &mut MemoryFill)

Visit MemoryFill.

fn visit_load_mut(&mut self, instr: &mut Load)

Visit Load.

fn visit_store_mut(&mut self, instr: &mut Store)

Visit Store.

fn visit_atomic_rmw_mut(&mut self, instr: &mut AtomicRmw)

Visit AtomicRmw.

fn visit_cmpxchg_mut(&mut self, instr: &mut Cmpxchg)

Visit Cmpxchg.

fn visit_atomic_notify_mut(&mut self, instr: &mut AtomicNotify)

Visit AtomicNotify.

fn visit_atomic_wait_mut(&mut self, instr: &mut AtomicWait)

Visit AtomicWait.

fn visit_atomic_fence_mut(&mut self, instr: &mut AtomicFence)

Visit AtomicFence.

fn visit_table_get_mut(&mut self, instr: &mut TableGet)

Visit TableGet.

fn visit_table_set_mut(&mut self, instr: &mut TableSet)

Visit TableSet.

fn visit_table_grow_mut(&mut self, instr: &mut TableGrow)

Visit TableGrow.

fn visit_table_size_mut(&mut self, instr: &mut TableSize)

Visit TableSize.

fn visit_table_fill_mut(&mut self, instr: &mut TableFill)

Visit TableFill.

fn visit_ref_null_mut(&mut self, instr: &mut RefNull)

Visit RefNull.

fn visit_ref_is_null_mut(&mut self, instr: &mut RefIsNull)

Visit RefIsNull.

fn visit_ref_func_mut(&mut self, instr: &mut RefFunc)

Visit RefFunc.

fn visit_ref_as_non_null_mut(&mut self, instr: &mut RefAsNonNull)

Visit RefAsNonNull.

fn visit_br_on_null_mut(&mut self, instr: &mut BrOnNull)

Visit BrOnNull.

fn visit_br_on_non_null_mut(&mut self, instr: &mut BrOnNonNull)

Visit BrOnNonNull.

fn visit_call_ref_mut(&mut self, instr: &mut CallRef)

Visit CallRef.

fn visit_return_call_ref_mut(&mut self, instr: &mut ReturnCallRef)

Visit ReturnCallRef.

fn visit_ref_i31_mut(&mut self, instr: &mut RefI31)

Visit RefI31.

fn visit_i31_get_s_mut(&mut self, instr: &mut I31GetS)

Visit I31GetS.

fn visit_i31_get_u_mut(&mut self, instr: &mut I31GetU)

Visit I31GetU.

fn visit_ref_test_mut(&mut self, instr: &mut RefTest)

Visit RefTest.

fn visit_ref_cast_mut(&mut self, instr: &mut RefCast)

Visit RefCast.

fn visit_br_on_cast_mut(&mut self, instr: &mut BrOnCast)

Visit BrOnCast.

fn visit_br_on_cast_fail_mut(&mut self, instr: &mut BrOnCastFail)

Visit BrOnCastFail.

fn visit_any_convert_extern_mut(&mut self, instr: &mut AnyConvertExtern)

Visit AnyConvertExtern.

fn visit_extern_convert_any_mut(&mut self, instr: &mut ExternConvertAny)

Visit ExternConvertAny.

fn visit_ref_eq_mut(&mut self, instr: &mut RefEq)

Visit RefEq.

fn visit_struct_new_mut(&mut self, instr: &mut StructNew)

Visit StructNew.

fn visit_struct_new_default_mut(&mut self, instr: &mut StructNewDefault)

Visit StructNewDefault.

fn visit_struct_get_mut(&mut self, instr: &mut StructGet)

Visit StructGet.

fn visit_struct_get_s_mut(&mut self, instr: &mut StructGetS)

Visit StructGetS.

fn visit_struct_get_u_mut(&mut self, instr: &mut StructGetU)

Visit StructGetU.

fn visit_struct_set_mut(&mut self, instr: &mut StructSet)

Visit StructSet.

fn visit_array_new_mut(&mut self, instr: &mut ArrayNew)

Visit ArrayNew.

fn visit_array_new_default_mut(&mut self, instr: &mut ArrayNewDefault)

Visit ArrayNewDefault.

fn visit_array_new_fixed_mut(&mut self, instr: &mut ArrayNewFixed)

Visit ArrayNewFixed.

fn visit_array_new_data_mut(&mut self, instr: &mut ArrayNewData)

Visit ArrayNewData.

fn visit_array_new_elem_mut(&mut self, instr: &mut ArrayNewElem)

Visit ArrayNewElem.

fn visit_array_get_mut(&mut self, instr: &mut ArrayGet)

Visit ArrayGet.

fn visit_array_get_s_mut(&mut self, instr: &mut ArrayGetS)

Visit ArrayGetS.

fn visit_array_get_u_mut(&mut self, instr: &mut ArrayGetU)

Visit ArrayGetU.

fn visit_array_set_mut(&mut self, instr: &mut ArraySet)

Visit ArraySet.

fn visit_array_len_mut(&mut self, instr: &mut ArrayLen)

Visit ArrayLen.

fn visit_array_fill_mut(&mut self, instr: &mut ArrayFill)

Visit ArrayFill.

fn visit_array_copy_mut(&mut self, instr: &mut ArrayCopy)

Visit ArrayCopy.

fn visit_array_init_data_mut(&mut self, instr: &mut ArrayInitData)

Visit ArrayInitData.

fn visit_array_init_elem_mut(&mut self, instr: &mut ArrayInitElem)

Visit ArrayInitElem.

fn visit_i64_add128_mut(&mut self, instr: &mut I64Add128)

Visit I64Add128.

fn visit_i64_sub128_mut(&mut self, instr: &mut I64Sub128)

Visit I64Sub128.

fn visit_i64_mul_wide_s_mut(&mut self, instr: &mut I64MulWideS)

Visit I64MulWideS.

fn visit_i64_mul_wide_u_mut(&mut self, instr: &mut I64MulWideU)

Visit I64MulWideU.

fn visit_v128_bitselect_mut(&mut self, instr: &mut V128Bitselect)

Visit V128Bitselect.

fn visit_i8x16_swizzle_mut(&mut self, instr: &mut I8x16Swizzle)

Visit I8x16Swizzle.

fn visit_i8x16_shuffle_mut(&mut self, instr: &mut I8x16Shuffle)

Visit I8x16Shuffle.

fn visit_load_simd_mut(&mut self, instr: &mut LoadSimd)

Visit LoadSimd.

fn visit_table_init_mut(&mut self, instr: &mut TableInit)

Visit TableInit.

fn visit_elem_drop_mut(&mut self, instr: &mut ElemDrop)

Visit ElemDrop.

fn visit_table_copy_mut(&mut self, instr: &mut TableCopy)

Visit TableCopy.

fn visit_return_call_mut(&mut self, instr: &mut ReturnCall)

Visit ReturnCall.

fn visit_return_call_indirect_mut(&mut self, instr: &mut ReturnCallIndirect)

Visit ReturnCallIndirect.

fn visit_try_table_mut(&mut self, instr: &mut TryTable)

Visit TryTable.

fn visit_throw_mut(&mut self, instr: &mut Throw)

Visit Throw.

fn visit_throw_ref_mut(&mut self, instr: &mut ThrowRef)

Visit ThrowRef.

fn visit_try_mut(&mut self, instr: &mut Try)

Visit Try.

fn visit_rethrow_mut(&mut self, instr: &mut Rethrow)

Visit Rethrow.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.