Struct walrus::BlockBuilder

pub struct BlockBuilder<'a> { /* fields omitted */ }

A builder returned by block-construction methods to build up block expressions over time.

Methods

impl<'_> BlockBuilder<'_>

pub fn expr(&mut self, expr: ExprId)

Pushes a new expression to be executed in the block we're building

pub fn id(&self) -> BlockId

Returns the id of the block that we're building