walrus

Struct LocalFunction

pub struct LocalFunction {
    pub args: Vec<LocalId>,
    pub instruction_mapping: Vec<(usize, InstrLocId)>,
    pub original_range: Option<Range<usize>>,
    /* private fields */
}

A function defined locally within the wasm module.

Fields
args: Vec<LocalId>

Arguments to this function, and the locals that they’re assigned to.

instruction_mapping: Vec<(usize, InstrLocId)>

InstrSeqId list mapping to original instruction

original_range: Option<Range<usize>>

Original function binary range.

Implementations

impl LocalFunction

pub fn ty(&self) -> TypeId

Get this function’s type.

pub fn entry_block(&self) -> InstrSeqId

Get the id of this function’s entry block.

pub fn block(&self, id: InstrSeqId) -> &InstrSeq

Get the block associated with the given id.

pub fn block_mut(&mut self, id: InstrSeqId) -> &mut InstrSeq

Get the block associated with the given id.

pub fn builder(&self) -> &FunctionBuilder

Get access to a FunctionBuilder to continue adding instructions to this function.

pub fn builder_mut(&mut self) -> &mut FunctionBuilder

Get access to a FunctionBuilder to continue adding instructions to this function.

pub fn size(&self) -> u64

Get the size of this function, in number of instructions.

pub fn is_const(&self) -> bool

Is this function’s body a constant instruction?

pub fn used_data_segments(&self) -> HashSet<Id<Data>, BuildIdHasher>

Collect the set of data segments that are used in this function via memory.init or data.drop instructions.