walrus

Struct ModuleExports

pub struct ModuleExports { /* private fields */ }

The set of exports in a module.

Implementations

impl ModuleExports

pub fn get(&self, id: ExportId) -> &Export

Gets a reference to an export given its id

pub fn get_mut(&mut self, id: ExportId) -> &mut Export

Gets a reference to an export given its id

pub fn delete(&mut self, id: ExportId)

Delete an export entry from this module.

pub fn iter(&self) -> impl Iterator<Item = &Export>

Get a shared reference to this module’s exports.

pub fn iter_mut(&mut self) -> impl Iterator<Item = &mut Export>

Get a mutable reference to this module’s exports.

pub fn add(&mut self, name: &str, item: impl Into<ExportItem>) -> ExportId

Add a new export to this module

pub fn get_exported_func(&self, f: FunctionId) -> Option<&Export>

Get a reference to a function export given its function id.

pub fn get_func(&self, name: impl AsRef<str>) -> Result<FunctionId>

Retrieve an exported function by export name

pub fn get_exported_table(&self, t: TableId) -> Option<&Export>

Get a reference to a table export given its table id.

pub fn get_exported_memory(&self, m: MemoryId) -> Option<&Export>

Get a reference to a memory export given its export id.

pub fn get_exported_global(&self, g: GlobalId) -> Option<&Export>

Get a reference to a global export given its global id.

pub fn remove(&mut self, name: impl AsRef<str>) -> Result<()>

Delete an export by name from this module.