walrus

Trait CustomSectionId

pub trait CustomSectionId {
    type CustomSection: ?Sized;
}

A common trait for custom section identifiers.

Used in the ModuleCustomSections::get family of methods to perform type conversions from dyn CustomSection trait objects into the concrete Self::CustomSection type instance.

You shouldn’t implement this yourself. Instead use TypedCustomSectionId<T> or UntypedCustomSectionId.

Required Associated Types

type CustomSection: ?Sized

The concrete custom section type that this id gets out of a ModuleCustomSections.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.